refactor(cmd): update cnosdb_tools to cnosdb-tools
