+++ cmd/cnosdb-tools/importer/importer.go
@@ -8,9 +8,9 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/cnosdb/cnosdb/cmd/cnosdb_tools/internal/errlist"
-	"github.com/cnosdb/cnosdb/cmd/cnosdb_tools/internal/shard"
-	"github.com/cnosdb/cnosdb/cmd/cnosdb_tools/server"
+	"github.com/cnosdb/cnosdb/cmd/cnosdb-tools/internal/errlist"
+	"github.com/cnosdb/cnosdb/cmd/cnosdb-tools/internal/shard"
+	"github.com/cnosdb/cnosdb/cmd/cnosdb-tools/server"
 	"github.com/cnosdb/cnosdb/meta"
 	"github.com/cnosdb/cnosdb/vend/db/tsdb"
 	"github.com/cnosdb/cnosdb/vend/db/tsdb/engine/tsm1"
