+++ cmd/cnosdb-tools/importer/series_writer.go
@@ -5,11 +5,11 @@ import (
 	"path/filepath"
 	"strconv"
 
+	"github.com/cnosdb/cnosdb/cmd/cnosdb-tools/internal/errlist"
 	"github.com/cnosdb/cnosdb/vend/db/models"
 	"github.com/cnosdb/cnosdb/vend/db/tsdb"
 	"github.com/cnosdb/cnosdb/vend/db/tsdb/engine/tsm1"
 	"github.com/cnosdb/cnosdb/vend/db/tsdb/index/tsi1"
-	"github.com/cnosdb/cnosdb/cmd/cnosdb_tools/internal/errlist"
 )
 
 type seriesWriter struct {
