+++ cmd/cnosdb-tools/internal/format/binary/common.go
