+++ cmd/cnosdb-tools/internal/format/binary/messagetype_string.go
