+++ cmd/cnosdb-tools/internal/format/binary/reader.go
@@ -6,9 +6,9 @@ import (
 	"fmt"
 	"io"
 
+	"github.com/cnosdb/cnosdb/cmd/cnosdb-tools/internal/tlv"
 	"github.com/cnosdb/cnosdb/vend/db/tsdb"
 	"github.com/cnosdb/cnosdb/vend/db/tsdb/engine/tsm1"
-	"github.com/cnosdb/cnosdb/cmd/cnosdb_tools/internal/tlv"
 )
 
 type Reader struct {
