+++ cmd/cnosdb-tools/internal/format/binary/reader_test.go
