+++ cmd/cnosdb-tools/internal/format/conflictwriter.go
