+++ cmd/cnosdb-tools/internal/shard/writer.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"strconv"
 
-	"github.com/cnosdb/cnosdb/cmd/cnosdb_tools/internal/errlist"
+	"github.com/cnosdb/cnosdb/cmd/cnosdb-tools/internal/errlist"
 	"github.com/cnosdb/cnosdb/vend/db/pkg/data/gen"
 	"github.com/cnosdb/cnosdb/vend/db/tsdb/engine/tsm1"
 )
