+++ cmd/cnosdb-tools/main.go
@@ -1,23 +1,23 @@
-// The cnosdb_tools command displays detailed information about CnosDB data files.
+// The cnosdb-tools command displays detailed information about CnosDB data files.
 package main
 
 import (
 	"fmt"
 
-	"github.com/cnosdb/cnosdb/cmd/cnosdb_tools/compact"
+	"github.com/cnosdb/cnosdb/cmd/cnosdb-tools/compact"
 
+	_ "github.com/cnosdb/cnosdb/cmd/cnosdb-tools/export"
+	_ "github.com/cnosdb/cnosdb/cmd/cnosdb-tools/generate/exec"
+	_ "github.com/cnosdb/cnosdb/cmd/cnosdb-tools/generate/init"
+	_ "github.com/cnosdb/cnosdb/cmd/cnosdb-tools/help"
+	_ "github.com/cnosdb/cnosdb/cmd/cnosdb-tools/importer"
+	_ "github.com/cnosdb/cnosdb/cmd/cnosdb-tools/server"
 	_ "github.com/cnosdb/cnosdb/cmd/cnosdb/run"
-	_ "github.com/cnosdb/cnosdb/cmd/cnosdb_tools/export"
-	_ "github.com/cnosdb/cnosdb/cmd/cnosdb_tools/generate/exec"
-	_ "github.com/cnosdb/cnosdb/cmd/cnosdb_tools/generate/init"
-	_ "github.com/cnosdb/cnosdb/cmd/cnosdb_tools/help"
-	_ "github.com/cnosdb/cnosdb/cmd/cnosdb_tools/importer"
-	_ "github.com/cnosdb/cnosdb/cmd/cnosdb_tools/server"
 	_ "github.com/cnosdb/cnosdb/meta"
 	_ "github.com/cnosdb/cnosdb/vend/db/tsdb"
 	_ "github.com/cnosdb/cnosdb/vend/db/tsdb/engine"
 
-	geninit "github.com/cnosdb/cnosdb/cmd/cnosdb_tools/generate/init"
+	geninit "github.com/cnosdb/cnosdb/cmd/cnosdb-tools/generate/init"
 	"github.com/spf13/cobra"
 )
 
@@ -39,7 +39,7 @@ func main() {
 
 func GetCommand() *cobra.Command {
 	c := &cobra.Command{
-		Use:  "cnosdb_tools",
+		Use:  "cnosdb-tools",
 		Long: "tools for managing and querying CnosDB data",
 		CompletionOptions: cobra.CompletionOptions{
 			DisableDefaultCmd:   true,
